refactor(utils): Phase 2 - Move book utilities to subdirectories
Organizational refactor - no functional changes:
- src/utils/book/book-metadata.ts (ISBN metadata utilities)
- src/utils/book/quality-scoring.ts (ISBNdb quality weights)
- src/utils/book/confidence.ts (confidence threshold utilities)
- src/utils/book/string-similarity.ts (Levenshtein distance)

Updated imports in 8 files:
- src/handlers/book-search.ts
- src/services/normalizers/*.ts (4 files)
- src/services/genre-normalizer.ts
- src/services/ai-scanner.ts
- tests/unit/string-similarity.test.ts

All 199 smoke tests passing.
Part of Phase 2: Medium-risk utility consolidation.
