+++ src/handlers/book-search.ts
@@ -12,7 +12,7 @@ import * as externalApis from '../services/external-apis'
 import { UnifiedCacheService } from '../services/unified-cache'
 import type { Env } from '../types/env'
 import { writeCacheMetrics } from '../utils/analytics/analytics'
-import { detectImageQuality } from '../utils/book-metadata'
+import { detectImageQuality } from '../utils/book/book-metadata'
 import { setCached } from '../utils/cache/cache'
 import { transformWorkToGoogleFormat } from '../utils/transform-work'
