+++ src/services/normalizers/openlibrary.ts
@@ -3,7 +3,7 @@
  */
 
 import type { AuthorDTO, EditionDTO, WorkDTO } from '../../types/canonical.js'
-import { getPlaceholderCover } from '../../utils/book-metadata.ts'
+import { getPlaceholderCover } from '../../utils/book/book-metadata.ts'
 import { extractYear } from '../../utils/date-utils.js'
 import { GenreNormalizer } from '../genre-normalizer.js'
