+++ src/utils/book/confidence.ts
