+++ src/utils/book/quality-scoring.ts
